fix: Move simulation setup into economy service
