+++ policyengine_api/libs/simulation_api.py
@@ -111,66 +111,3 @@ class SimulationAPI:
         """
 
         return self.execution_client.get_execution(name=id)
-
-    def _setup_sim_options(
-        self,
-        country_id: str,
-        reform_policy: Annotated[str, "String-formatted JSON"],
-        baseline_policy: Annotated[str, "String-formatted JSON"],
-        region: str,
-        dataset: str | None,
-        time_period: str,
-        scope: Literal["macro", "household"] = "macro",
-        include_cliffs: bool = False,
-        model_version: str | None = None,
-        data_version: str | None = None,
-    ) -> dict[str, Any]:
-        """
-        Set up the simulation options for the simulation API job.
-        """
-
-        return {
-            "country": country_id,
-            "scope": scope,
-            "reform": json.loads(reform_policy),
-            "baseline": json.loads(baseline_policy),
-            "time_period": time_period,
-            "include_cliffs": include_cliffs,
-            "region": self._setup_region(country_id=country_id, region=region),
-            "data": self._setup_data(
-                dataset=dataset, country_id=country_id, region=region
-            ),
-            "model_version": model_version,
-            "data_version": data_version,
-        }
-
-    def _setup_region(self, country_id: str, region: str) -> str:
-        """
-        Convert API v1 'region' option to API v2-compatible 'region' option.
-        """
-
-        # For US, states must be prefixed with 'state/'
-        if country_id == "us" and region != "us":
-            return "state/" + region
-
-        return region
-
-    def _setup_data(
-        self, dataset: str | None, country_id: str, region: str
-    ) -> str | None:
-        """
-        Take API v1 'data' string literals, which reference a dataset name,
-        and convert to relevant GCP filepath. In future, this should be
-        redone to use a more robust method of accessing datasets.
-        """
-
-        # Enhanced CPS runs must reference ECPS dataset in Google Cloud bucket
-        if dataset == "enhanced_cps":
-            return "gs://policyengine-us-data/enhanced_cps_2024.h5"
-
-        # US state-level simulations must reference pooled CPS dataset
-        if country_id == "us" and region != "us":
-            return "gs://policyengine-us-data/pooled_3_year_cps_2023.h5"
-
-        # All others receive no sim API 'data' arg
-        return None
